move borderless textarea to UI
