+++ apps/desktop/src/lib/navigation/ProjectsPopup.svelte
@@ -2,9 +2,9 @@
 	import { ProjectService } from '$lib/backend/projects';
 	import ScrollableContainer from '$lib/scroll/ScrollableContainer.svelte';
 	import { getContext } from '$lib/utils/context';
-	import { resizeObserver } from '$lib/utils/resizeObserver';
 	import Icon from '@gitbutler/ui/Icon.svelte';
 	import { portal } from '@gitbutler/ui/utils/portal';
+	import { resizeObserver } from '@gitbutler/ui/utils/resizeObserver';
 	import type iconsJson from '@gitbutler/ui/data/icons.json';
 	import { goto } from '$app/navigation';
 	import { page } from '$app/stores';
